Issue #OS-76 Refactor commons package
