--- java/middleware-commons/src/main/java/io/opensaber/registry/middleware/transform/commons/Constants.java
@@ -1,13 +1,0 @@
-package io.opensaber.registry.middleware.transform.commons;
-
-public class Constants {
-
-    public static class JsonldConstants {
-        public static final String CONTEXT = "@context";
-        public static final String ID = "@id";
-        public static final String TYPE = "@type";
-        public static final String VALUE = "@value";
-        public static final String GRAPH = "@graph";
-    }
-
-}
