+++ java/registry-interceptor/src/main/java/io/opensaber/registry/interceptor/RDFConversionInterceptor.java
@@ -18,7 +18,7 @@ import io.opensaber.registry.interceptor.request.transform.RequestTransformFacto
 import io.opensaber.registry.middleware.Middleware;
 import io.opensaber.registry.middleware.MiddlewareHaltException;
 import io.opensaber.registry.middleware.impl.RDFConverter;
-import io.opensaber.registry.middleware.transform.commons.Data;
+import io.opensaber.registry.middleware.transform.Data;
 import io.opensaber.registry.middleware.util.Constants;
 
 @Component
