+++ java/registry-interceptor/src/main/java/io/opensaber/registry/interceptor/request/transform/JsonldToLdRequestTransformer.java
@@ -6,9 +6,9 @@ import java.util.List;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
-import io.opensaber.registry.middleware.transform.commons.Data;
-import io.opensaber.registry.middleware.transform.commons.ITransformer;
-import io.opensaber.registry.middleware.transform.commons.TransformationException;
+import io.opensaber.registry.middleware.transform.Data;
+import io.opensaber.registry.middleware.transform.ITransformer;
+import io.opensaber.registry.middleware.transform.TransformationException;
 
 
 public class JsonldToLdRequestTransformer implements ITransformer<Object> {
