+++ java/registry/src/main/java/io/opensaber/registry/transformation/JsonToLdTransformer.java
@@ -6,11 +6,11 @@ import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.fasterxml.jackson.databind.node.JsonNodeFactory;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 
-import io.opensaber.registry.middleware.transform.commons.Data;
-import io.opensaber.registry.middleware.transform.commons.ErrorCode;
-import io.opensaber.registry.middleware.transform.commons.ITransformer;
-import io.opensaber.registry.middleware.transform.commons.TransformationException;
-import io.opensaber.registry.middleware.transform.commons.Constants.JsonldConstants;
+import io.opensaber.registry.middleware.transform.Data;
+import io.opensaber.registry.middleware.transform.ErrorCode;
+import io.opensaber.registry.middleware.transform.ITransformer;
+import io.opensaber.registry.middleware.transform.TransformationException;
+import io.opensaber.registry.middleware.util.Constants.JsonldConstants;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -31,7 +31,7 @@ public class JsonToLdTransformer implements ITransformer<Object> {
 	public Data<Object> transform(Data<Object> data) throws TransformationException {
 		try {
 			ObjectMapper mapper = new ObjectMapper();
-			ObjectNode input = (ObjectNode) mapper.readTree(data.getData().toString());			
+			ObjectNode input = (ObjectNode) mapper.readTree(data.getData().toString());
 			JsonNode jsonNode = getconstructedJson(input, keysToPurge);
 			return new Data<>(jsonNode);
 		} catch (Exception ex) {
@@ -45,7 +45,7 @@ public class JsonToLdTransformer implements ITransformer<Object> {
 
 		setPurgeData(keysToPurge);
 		ArrayNode arrayNode = JsonNodeFactory.instance.arrayNode();
-		for(JsonNode graphNode : rootDataNode.path(JsonldConstants.GRAPH)){
+		for (JsonNode graphNode : rootDataNode.path(JsonldConstants.GRAPH)) {
 			ObjectNode rootNode = addRootTypeNode(graphNode);
 			if (keysToPurge.size() != 0)
 				purgedKeys(rootNode);
@@ -69,23 +69,22 @@ public class JsonToLdTransformer implements ITransformer<Object> {
 			if (keysToPurge.contains(entry.getKey())) {
 				removeKeyNames.add(entry.getKey());
 			} else {
-				if(entry.getValue().isValueNode()){
-					if(entry.getValue().toString().contains(prefix))
-						replaceText(node,entry.getKey(),entry.getValue().asText().replace(prefix, "") );
-					
-				}else if (entry.getValue().isArray()) {
+				if (entry.getValue().isValueNode()) {
+					if (entry.getValue().toString().contains(prefix))
+						replaceText(node, entry.getKey(), entry.getValue().asText().replace(prefix, ""));
+
+				} else if (entry.getValue().isArray()) {
 					for (int i = 0; i < entry.getValue().size(); i++) {
-						if (entry.getValue().get(i).isObject()) 
+						if (entry.getValue().get(i).isObject())
 							purgedKeys((ObjectNode) entry.getValue().get(i));
 					}
-				} else if (entry.getValue().isObject()) {					
+				} else if (entry.getValue().isObject()) {
 					purgedKeys((ObjectNode) entry.getValue());
 				}
 			}
 		});
 		node.remove(removeKeyNames);
 	}
-	
 
 	public static void replaceText(JsonNode parent, String fieldName, String newText) {
 		if (parent.has(fieldName)) {
@@ -98,11 +97,9 @@ public class JsonToLdTransformer implements ITransformer<Object> {
 		this.keysToPurge = keyToPruge;
 
 	}
-	
-	private void setPrefix(String prefix){
+
+	private void setPrefix(String prefix) {
 		this.prefix = prefix + SEPERATOR;
 	}
-
-	
 
 }
